@@ -1,55 +1,67 @@
 {
-    "name": "bog-npp",
-    "version": "0.0.1",
-    "main": "index.js",
-    "scripts": {
-        "dev": "next",
-        "hardcode": "hardcode email/dev/** -p email/dev/ -o email/import/",
-        "build": "next build",
-        "start": "next start",
-        "lint": "eslint --ext .js --ext .jsx ."
-    },
-    "dependencies": {
-        "@fortawesome/fontawesome-svg-core": "^1.2.25",
-        "@fortawesome/free-solid-svg-icons": "^5.11.2",
-        "@fortawesome/react-fontawesome": "^0.1.4",
-        "@zeit/next-css": "^1.0.1",
-        "bcryptjs": "^2.4.3",
-        "bootstrap": "^4.3.1",
-        "email-templates": "^6.0.3",
-        "font-awesome": "^4.7.0",
-        "formik": "^2.0.3",
-        "hardcode": "^1.4.0",
-        "isomorphic-unfetch": "^3.0.0",
-        "js-cookie": "^2.2.1",
-        "jsonwebtoken": "^8.5.1",
-        "lodash": "^4.17.19",
-        "moment": "^2.24.0",
-        "mongodb": "^3.6.2",
-        "mongoose": "^5.10.6",
-        "next": "^9.3.2",
-        "next-cookies": "^2.0.1",
-        "next-images": "^1.2.0",
-        "prop-types": "^15.7.2",
-        "pug": "^3.0.1",
-        "react": "^16.9.0",
-        "react-dom": "^16.9.0",
-        "react-redux": "^7.1.1",
-        "react-transition-group": "^4.3.0",
-        "reactstrap": "^8.2.0",
-        "redux": "^4.0.4",
-        "redux-devtools-extension": "^2.13.8",
-        "redux-thunk": "^2.3.0",
-        "uuid": "^3.3.3",
-        "yup": "^0.27.0"
-    },
-    "devDependencies": {
-        "babel-eslint": "^10.0.3",
-        "eslint": "^6.4.0",
-        "eslint-config-airbnb": "^18.0.1",
-        "eslint-plugin-import": "^2.18.2",
-        "eslint-plugin-jsx-a11y": "^6.2.3",
-        "eslint-plugin-react": "^7.14.3",
-        "eslint-plugin-react-hooks": "^1.7.0"
-    }
+  "name": "bog-npp",
+  "version": "0.0.1",
+  "main": "index.js",
+  "scripts": {
+    "dev": "next",
+    "hardcode": "hardcode server/email/dev/** -p server/email/dev/ -o server/email/import/",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint",
+    "lint:fix": "next lint --fix",
+    "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,css,scss}\"",
+    "prepare": "husky install"
+  },
+  "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^1.3.0",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/react-fontawesome": "^0.1.17",
+    "@redux-devtools/extension": "^3.2.2",
+    "axios": "^0.26.1",
+    "bcrypt": "^5.0.1",
+    "bcryptjs": "^2.4.3",
+    "bootstrap": "^4.3.1",
+    "clsx": "^1.1.1",
+    "email-templates": "^6.1.1",
+    "font-awesome": "^4.7.0",
+    "formik": "^2.2.9",
+    "hardcode": "^1.4.0",
+    "iron-session": "^6.0.5",
+    "lodash": "^4.17.21",
+    "moment": "^2.29.1",
+    "mongodb": "^4.4.1",
+    "mongoose": "^6.2.5",
+    "next": "^12.1.0",
+    "next-images": "^1.8.4",
+    "next-redux-wrapper": "^7.0.5",
+    "prop-types": "^15.8.1",
+    "pug": "^3.0.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.6",
+    "react-transition-group": "^4.4.2",
+    "reactstrap": "^8.10.1",
+    "redux": "^4.1.2",
+    "redux-thunk": "^2.4.1",
+    "swr": "^1.2.2",
+    "uuid": "^8.3.2",
+    "yup": "^0.32.11"
+  },
+  "devDependencies": {
+    "eslint": "8.10.0",
+    "eslint-config-next": "12.1.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "husky": "^7.0.4",
+    "lint-staged": "^12.3.5",
+    "prettier": "^2.5.1"
+  },
+  "lint-staged": {
+    "*.{js,jsx}": [
+      "eslint --fix"
+    ],
+    "*.{js,jsx,css}": [
+      "prettier --write"
+    ]
+  }
 }
