@@ -1,4 +1,3 @@
-const withCSS = require('@zeit/next-css');
-const withImages = require('next-images');
+const withImages = require("next-images");
 
-module.exports = withImages(withCSS());
+module.exports = withImages();
