@@ -1,20 +1,21 @@
-import mongoose from 'mongoose';
-import config from '../../config';
+import mongoose from "mongoose";
+
+const prod = process.env.NODE_ENV === "production";
+const dbUrl = prod
+  ? process.env.MONGO_DB || process.env.MONGO_DEV_DB
+  : process.env.MONGO_DEV_DB || "mongodb://localhost:27017";
+const dbName = process.env.DB_NAME || "Applications";
 
 export default async () => {
   if (mongoose.connections[0].readyState) return;
 
-  await mongoose.connect(config.dbUrl, {
-    // avoids deprecated functionality
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-    useFindAndModify: false,
-    useCreateIndex: true,
-    dbName: config.dbName,
-  })
-    .catch((e) => {
-      console.error('Error connecting to database.');
-
-      throw e;
+  try {
+    await mongoose.connect(dbUrl, {
+      dbName,
     });
+  } catch (error) {
+    console.error("Error connecting to database.");
+
+    throw error;
+  }
 };
