@@ -1,6 +1,6 @@
-import pug from 'pug';
-import Email from 'email-templates';
-import emailFolder from '../../email/import/index';
+import pug from "pug";
+import Email from "email-templates";
+import emailFolder from "../email/import/index";
 
 const fromAddress = '"GT Bits of Good" <hello@bitsofgood.org>';
 
@@ -16,12 +16,13 @@ const transportConfig = {
 
 function _render(email, view, locals) {
   return new Promise((resolve, reject) => {
-    const viewParts = view.split('/');
+    const viewParts = view.split("/");
     const pugTemplate = emailFolder[viewParts[0]][`${viewParts[1]}.pug`];
-    if (typeof pugTemplate === 'undefined') {
+    if (typeof pugTemplate === "undefined") {
       resolve();
     }
-    email.juiceResources(pug.render(pugTemplate, { filename: view, ...locals }))
+    email
+      .juiceResources(pug.render(pugTemplate, { filename: view, ...locals }))
       .then((juicedHTML) => {
         resolve(juicedHTML);
       })
@@ -39,18 +40,16 @@ export const sendEmail = (options) => {
     send: process.env.MAIL_HOST != null,
     juice: true,
     juiceResources: {
-      extraCss: emailFolder[options.template]['style.css'],
+      extraCss: emailFolder[options.template]["style.css"],
     },
     render: (view, locals) => _render(email, view, locals),
   });
 
-  return email.send(
-    {
-      template: options.template,
-      message: {
-        to: options.to,
-      },
-      locals: options.locals,
+  return email.send({
+    template: options.template,
+    message: {
+      to: options.to,
     },
-  );
+    locals: options.locals,
+  });
 };
